+++ baggage/baggage.go
@@ -17,7 +17,7 @@ package baggage // import "go.opentelemetry.io/otel/baggage"
 import (
 	"context"
 
-	"go.opentelemetry.io/otel/internal/baggage"
+	"go.opentelemetry.io/otel/baggage/updatable"
 	"go.opentelemetry.io/otel/label"
 )
 
@@ -27,7 +27,7 @@ func Set(ctx context.Context) label.Set {
 	// the functional elements of the label.Set. These should be unified so
 	// this conversion is unnecessary and there is no performance hit calling
 	// this.
-	m := baggage.MapFromContext(ctx)
+	m := updatable.MapFromContext(ctx)
 	values := make([]label.KeyValue, 0, m.Len())
 	m.Foreach(func(kv label.KeyValue) bool {
 		values = append(values, kv)
@@ -40,28 +40,28 @@ func Set(ctx context.Context) label.Set {
 // value is set, the returned label.Value will be an uninitialized zero-value
 // with type INVALID.
 func Value(ctx context.Context, key label.Key) label.Value {
-	v, _ := baggage.MapFromContext(ctx).Value(key)
+	v, _ := updatable.MapFromContext(ctx).Value(key)
 	return v
 }
 
 // ContextWithValues returns a copy of parent with pairs updated in the baggage.
 func ContextWithValues(parent context.Context, pairs ...label.KeyValue) context.Context {
-	m := baggage.MapFromContext(parent).Apply(baggage.MapUpdate{
+	m := updatable.MapFromContext(parent).Apply(updatable.MapUpdate{
 		MultiKV: pairs,
 	})
-	return baggage.ContextWithMap(parent, m)
+	return updatable.ContextWithMap(parent, m)
 }
 
 // ContextWithoutValues returns a copy of parent in which the values related
 // to keys have been removed from the baggage.
 func ContextWithoutValues(parent context.Context, keys ...label.Key) context.Context {
-	m := baggage.MapFromContext(parent).Apply(baggage.MapUpdate{
+	m := updatable.MapFromContext(parent).Apply(updatable.MapUpdate{
 		DropMultiK: keys,
 	})
-	return baggage.ContextWithMap(parent, m)
+	return updatable.ContextWithMap(parent, m)
 }
 
 // ContextWithEmpty returns a copy of parent without baggage.
 func ContextWithEmpty(parent context.Context) context.Context {
-	return baggage.ContextWithNoCorrelationData(parent)
+	return updatable.ContextWithNoCorrelationData(parent)
 }