+++ bridge/opentracing/bridge.go
@@ -26,9 +26,9 @@ import (
 	otlog "github.com/opentracing/opentracing-go/log"
 
 	"go.opentelemetry.io/otel"
+	"go.opentelemetry.io/otel/baggage/updatable"
 	"go.opentelemetry.io/otel/bridge/opentracing/migration"
 	"go.opentelemetry.io/otel/codes"
-	"go.opentelemetry.io/otel/internal/baggage"
 	"go.opentelemetry.io/otel/internal/trace/noop"
 	otelparent "go.opentelemetry.io/otel/internal/trace/parent"
 	"go.opentelemetry.io/otel/label"
@@ -37,7 +37,7 @@ import (
 )
 
 type bridgeSpanContext struct {
-	baggageItems    baggage.Map
+	baggageItems    updatable.Map
 	otelSpanContext trace.SpanContext
 }
 
@@ -45,7 +45,7 @@ var _ ot.SpanContext = &bridgeSpanContext{}
 
 func newBridgeSpanContext(otelSpanContext trace.SpanContext, parentOtSpanContext ot.SpanContext) *bridgeSpanContext {
 	bCtx := &bridgeSpanContext{
-		baggageItems:    baggage.NewEmptyMap(),
+		baggageItems:    updatable.NewEmptyMap(),
 		otelSpanContext: otelSpanContext,
 	}
 	if parentOtSpanContext != nil {
@@ -65,7 +65,7 @@ func (c *bridgeSpanContext) ForeachBaggageItem(handler func(k, v string) bool) {
 
 func (c *bridgeSpanContext) setBaggageItem(restrictedKey, value string) {
 	crk := http.CanonicalHeaderKey(restrictedKey)
-	c.baggageItems = c.baggageItems.Apply(baggage.MapUpdate{SingleKV: label.String(crk, value)})
+	c.baggageItems = c.baggageItems.Apply(updatable.MapUpdate{SingleKV: label.String(crk, value)})
 }
 
 func (c *bridgeSpanContext) baggageItem(restrictedKey string) string {
@@ -333,8 +333,8 @@ func (t *BridgeTracer) SetTextMapPropagator(propagator propagation.TextMapPropag
 }
 
 func (t *BridgeTracer) NewHookedContext(ctx context.Context) context.Context {
-	ctx = baggage.ContextWithSetHook(ctx, t.baggageSetHook)
-	ctx = baggage.ContextWithGetHook(ctx, t.baggageGetHook)
+	ctx = updatable.ContextWithSetHook(ctx, t.baggageSetHook)
+	ctx = updatable.ContextWithGetHook(ctx, t.baggageGetHook)
 	return ctx
 }
 
@@ -352,8 +352,8 @@ func (t *BridgeTracer) baggageSetHook(ctx context.Context) context.Context {
 	// we clear the context only to avoid calling a get hook
 	// during MapFromContext, but otherwise we don't change the
 	// context, so we don't care about the old hooks.
-	clearCtx, _, _ := baggage.ContextWithNoHooks(ctx)
-	m := baggage.MapFromContext(clearCtx)
+	clearCtx, _, _ := updatable.ContextWithNoHooks(ctx)
+	m := updatable.MapFromContext(clearCtx)
 	m.Foreach(func(kv label.KeyValue) bool {
 		bSpan.setBaggageItemOnly(string(kv.Key), kv.Value.Emit())
 		return true
@@ -361,7 +361,7 @@ func (t *BridgeTracer) baggageSetHook(ctx context.Context) context.Context {
 	return ctx
 }
 
-func (t *BridgeTracer) baggageGetHook(ctx context.Context, m baggage.Map) baggage.Map {
+func (t *BridgeTracer) baggageGetHook(ctx context.Context, m updatable.Map) updatable.Map {
 	span := ot.SpanFromContext(ctx)
 	if span == nil {
 		t.warningHandler("No active OpenTracing span, can not propagate the baggage items from OpenTracing span context\n")
@@ -380,7 +380,7 @@ func (t *BridgeTracer) baggageGetHook(ctx context.Context, m baggage.Map) baggag
 	for k, v := range items {
 		kv = append(kv, label.String(k, v))
 	}
-	return m.Apply(baggage.MapUpdate{MultiKV: kv})
+	return m.Apply(updatable.MapUpdate{MultiKV: kv})
 }
 
 // StartSpan is a part of the implementation of the OpenTracing Tracer
@@ -619,7 +619,7 @@ func (t *BridgeTracer) Inject(sm ot.SpanContext, format interface{}, carrier int
 		sc:   bridgeSC.otelSpanContext,
 	}
 	ctx := trace.ContextWithSpan(context.Background(), fs)
-	ctx = baggage.ContextWithMap(ctx, bridgeSC.baggageItems)
+	ctx = updatable.ContextWithMap(ctx, bridgeSC.baggageItems)
 	t.getPropagator().Inject(ctx, header)
 	return nil
 }
@@ -638,7 +638,7 @@ func (t *BridgeTracer) Extract(format interface{}, carrier interface{}) (ot.Span
 	}
 	header := http.Header(hhcarrier)
 	ctx := t.getPropagator().Extract(context.Background(), header)
-	baggage := baggage.MapFromContext(ctx)
+	baggage := updatable.MapFromContext(ctx)
 	otelSC, _, _ := otelparent.GetSpanContextAndLinks(ctx, false)
 	bridgeSC := &bridgeSpanContext{
 		baggageItems:    baggage,