+++ CHANGELOG.md
@@ -11,6 +11,7 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 ### Changed
 
 - Reverse order in which `Resource` attributes are merged, per change in spec. (#1501)
+- Rename `internal/baggage` to `baggage/updatable`.
 
 ## [0.16.0] - 2020-01-13